refactor: divide common hash key and schemas
