+++ core/storage/src/lib.rs
@@ -1,13 +1,14 @@
 #![allow(clippy::mutable_key_type)]
 pub mod adapter;
 mod cache;
+mod hash_key;
+mod schema;
 #[cfg(test)]
 mod tests;
 
 use std::collections::{HashMap, HashSet};
 use std::convert::From;
 use std::error::Error;
-use std::str::FromStr;
 use std::sync::Arc;
 
 use arc_swap::ArcSwap;
@@ -29,6 +30,12 @@ use protocol::{
 };
 
 use crate::cache::StorageCache;
+use crate::hash_key::{BlockKey, CommonHashKey, CommonPrefix};
+use crate::schema::{
+    BlockHashNumberSchema, BlockHeaderSchema, BlockSchema, EvmCodeAddressSchema, EvmCodeSchema,
+    LatestBlockSchema, ReceiptBytesSchema, ReceiptSchema, TransactionBytesSchema,
+    TransactionSchema, TxHashNumberSchema, LatestProofSchema, CrossChainRecordSchema
+};
 
 const BATCH_VALUE_DECODE_NUMBER: usize = 1000;
 
@@ -78,21 +85,6 @@ macro_rules! ensure_get {
         let opt = get!($self_, $key, $schema)?;
         opt.ok_or_else(|| StorageError::GetNone($key.to_string()))?
     }};
-}
-
-macro_rules! impl_storage_schema_for {
-    ($name: ident, $key: ident, $val: ident, $category: ident) => {
-        pub struct $name;
-
-        impl StorageSchema for $name {
-            type Key = $key;
-            type Value = $val;
-
-            fn category() -> StorageCategory {
-                StorageCategory::$category
-            }
-        }
-    };
 }
 
 #[derive(Debug)]
@@ -187,155 +179,6 @@ impl<Adapter: StorageAdapter> ImplStorage<Adapter> {
         Ok(())
     }
 }
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct CommonPrefix {
-    block_height: [u8; 8], // BigEndian
-}
-
-impl CommonPrefix {
-    pub fn new(block_height: u64) -> Self {
-        CommonPrefix {
-            block_height: block_height.to_be_bytes(),
-        }
-    }
-
-    pub fn len() -> usize {
-        8
-    }
-
-    pub fn height(self) -> u64 {
-        u64::from_be_bytes(self.block_height)
-    }
-
-    pub fn make_hash_key(self, hash: &Hash) -> [u8; 40] {
-        debug_assert!(hash.as_bytes().len() == 32);
-
-        let mut key = [0u8; 40];
-        key[0..8].copy_from_slice(&self.block_height);
-        key[8..40].copy_from_slice(&hash.as_bytes()[..32]);
-
-        key
-    }
-}
-
-impl AsRef<[u8]> for CommonPrefix {
-    fn as_ref(&self) -> &[u8] {
-        &self.block_height
-    }
-}
-
-impl From<&[u8]> for CommonPrefix {
-    fn from(bytes: &[u8]) -> CommonPrefix {
-        debug_assert!(bytes.len() >= 8);
-
-        let mut h_buf = [0u8; 8];
-        h_buf.copy_from_slice(&bytes[0..8]);
-
-        CommonPrefix {
-            block_height: h_buf,
-        }
-    }
-}
-
-impl ProtocolCodec for CommonPrefix {
-    fn encode(&self) -> ProtocolResult<Bytes> {
-        Ok(Bytes::copy_from_slice(&self.block_height))
-    }
-
-    fn decode<B: AsRef<[u8]>>(bytes: B) -> ProtocolResult<Self> {
-        Ok(CommonPrefix::from(&bytes.as_ref()[..8]))
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct CommonHashKey {
-    prefix: CommonPrefix,
-    hash:   Hash,
-}
-
-impl CommonHashKey {
-    pub fn new(block_height: u64, hash: Hash) -> Self {
-        CommonHashKey {
-            prefix: CommonPrefix::new(block_height),
-            hash,
-        }
-    }
-
-    pub fn height(&self) -> u64 {
-        self.prefix.height()
-    }
-
-    pub fn hash(&self) -> &Hash {
-        &self.hash
-    }
-}
-
-impl ProtocolCodec for CommonHashKey {
-    fn encode(&self) -> ProtocolResult<Bytes> {
-        Ok(Bytes::copy_from_slice(
-            &self.prefix.make_hash_key(&self.hash),
-        ))
-    }
-
-    fn decode<B: AsRef<[u8]>>(bytes: B) -> ProtocolResult<Self> {
-        let mut bytes = bytes.as_ref().to_vec();
-        debug_assert!(bytes.len() >= CommonPrefix::len());
-
-        let prefix = CommonPrefix::from(&bytes[0..CommonPrefix::len()]);
-        let hash = Hash::from_slice(&bytes.split_off(CommonPrefix::len()));
-
-        Ok(CommonHashKey { prefix, hash })
-    }
-}
-
-impl ToString for CommonHashKey {
-    fn to_string(&self) -> String {
-        format!("{}:{}", self.prefix.height(), self.hash)
-    }
-}
-
-impl FromStr for CommonHashKey {
-    type Err = ();
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let parts = s.split(':').collect::<Vec<_>>();
-        debug_assert!(parts.len() == 2);
-
-        let height = parts[0].parse::<u64>().map_err(|_| ())?;
-
-        let hash = Hasher::digest(parts[1].as_bytes());
-
-        Ok(CommonHashKey::new(height, hash))
-    }
-}
-
-pub type BlockKey = CommonPrefix;
-
-impl_storage_schema_for!(
-    TransactionSchema,
-    CommonHashKey,
-    SignedTransaction,
-    SignedTransaction
-);
-impl_storage_schema_for!(
-    TransactionBytesSchema,
-    CommonHashKey,
-    DBBytes,
-    SignedTransaction
-);
-impl_storage_schema_for!(BlockSchema, BlockKey, Block, Block);
-impl_storage_schema_for!(BlockHeaderSchema, BlockKey, Header, BlockHeader);
-impl_storage_schema_for!(BlockHashNumberSchema, Hash, u64, HashHeight);
-impl_storage_schema_for!(ReceiptSchema, CommonHashKey, Receipt, Receipt);
-impl_storage_schema_for!(ReceiptBytesSchema, CommonHashKey, DBBytes, Receipt);
-impl_storage_schema_for!(TxHashNumberSchema, Hash, u64, HashHeight);
-impl_storage_schema_for!(LatestBlockSchema, Hash, Block, Block);
-impl_storage_schema_for!(LatestProofSchema, Hash, Proof, Block);
-impl_storage_schema_for!(OverlordWalSchema, Hash, Bytes, Wal);
-impl_storage_schema_for!(EvmCodeSchema, Hash, Bytes, Code);
-impl_storage_schema_for!(EvmCodeAddressSchema, Hash, Hash, Code);
-impl_storage_schema_for!(CrossChainRecordSchema, Hash, HashWithDirection, CrossChain);
 
 #[async_trait]
 impl<Adapter: StorageAdapter> CommonStorage for ImplStorage<Adapter> {
@@ -487,10 +330,10 @@ impl<Adapter: StorageAdapter> Storage for ImplStorage<Adapter> {
                 // Note: fix clippy::suspicious_else_formatting
                 if key.height() != block_height {
                     break;
-                } else if !set.contains(&key.hash) {
+                } else if !set.contains(key.hash()) {
                     continue;
                 } else {
-                    found.push((key.hash, stx_bytes));
+                    found.push((*key.hash(), stx_bytes));
                     count -= 1;
                 }
             }
@@ -656,10 +499,10 @@ impl<Adapter: StorageAdapter> Storage for ImplStorage<Adapter> {
                 // Note: fix clippy::suspicious_else_formatting
                 if key.height() != block_height {
                     break;
-                } else if !set.contains(&key.hash) {
+                } else if !set.contains(key.hash()) {
                     continue;
                 } else {
-                    found.push((key.hash, stx_bytes));
+                    found.push((*key.hash(), stx_bytes));
                     count -= 1;
                 }
             }
